+++ coupon-service/data/jpa/src/main/kotlin/com/hyoseok/config/jpa/JpaConfig.kt
@@ -17,7 +17,7 @@ import javax.sql.DataSource
 @Configuration
 @EnableTransactionManagement
 @EnableJpaRepositories(
-    basePackages = ["com.hyoseok.repository"],
+    basePackages = ["com.hyoseok.*.repository"],
     entityManagerFactoryRef = "entityManagerFactory",
     transactionManagerRef = "jpaTransactionManager",
 )
