+++ coupon-service/data/jpa/src/main/kotlin/com/hyoseok/coupon/repository/CouponIssuedFailJpaRepository.kt
@@ -1,0 +1,6 @@
+package com.hyoseok.coupon.repository
+
+import com.hyoseok.coupon.entity.CouponIssuedFailEntity
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface CouponIssuedFailJpaRepository : JpaRepository<CouponIssuedFailEntity, Long>
