+++ coupon-service/data/jpa/src/main/kotlin/com/hyoseok/coupon/repository/CouponIssuedJpaRepository.kt
@@ -1,0 +1,6 @@
+package com.hyoseok.coupon.repository
+
+import com.hyoseok.coupon.entity.CouponIssuedEntity
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface CouponIssuedJpaRepository : JpaRepository<CouponIssuedEntity, Long>
