+++ coupon-service/data/jpa/src/main/kotlin/com/hyoseok/coupon/repository/CouponJpaRepository.kt
@@ -1,0 +1,6 @@
+package com.hyoseok.coupon.repository
+
+import com.hyoseok.coupon.entity.CouponEntity
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface CouponJpaRepository : JpaRepository<CouponEntity, Long>
