+++ coupon-service/data/jpa/src/main/kotlin/com/hyoseok/member/repository/MemberJpaRepository.kt
@@ -1,0 +1,6 @@
+package com.hyoseok.member.repository
+
+import com.hyoseok.member.entity.MemberEntity
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface MemberJpaRepository : JpaRepository<MemberEntity, Long>
