--- coupon-service/data/jpa/src/main/kotlin/com/hyoseok/repository/coupon/CouponIssuedFailJpaRepository.kt
@@ -1,6 +1,0 @@
-package com.hyoseok.repository.coupon
-
-import com.hyoseok.entity.coupon.CouponIssuedFailEntity
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface CouponIssuedFailJpaRepository : JpaRepository<CouponIssuedFailEntity, Long>
