--- coupon-service/data/jpa/src/main/kotlin/com/hyoseok/repository/coupon/CouponIssuedJpaRepository.kt
@@ -1,6 +1,0 @@
-package com.hyoseok.repository.coupon
-
-import com.hyoseok.entity.coupon.CouponIssuedEntity
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface CouponIssuedJpaRepository : JpaRepository<CouponIssuedEntity, Long>
