--- coupon-service/data/jpa/src/main/kotlin/com/hyoseok/repository/coupon/CouponJpaRepository.kt
@@ -1,6 +1,0 @@
-package com.hyoseok.repository.coupon
-
-import com.hyoseok.entity.coupon.CouponEntity
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface CouponJpaRepository : JpaRepository<CouponEntity, Long>
