--- coupon-service/data/jpa/src/main/kotlin/com/hyoseok/repository/member/MemberJpaRepository.kt
@@ -1,6 +1,0 @@
-package com.hyoseok.repository.member
-
-import com.hyoseok.entity.member.MemberEntity
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface MemberJpaRepository : JpaRepository<MemberEntity, Long>
